upped the memory limit

+++ scripts/submit_clustal.pl
@@ -24,7 +24,7 @@ use Getopt::Long;
 use DBI;
 use strict;
 
-my ($debug,$store,$species,$wb,$user,$pass,$test,$ace_database);
+my ($debug,$store,$species,$wb,$user,$pass,$test,$ace_database,$dontclean);
 GetOptions(
  'debug=s'   => \$debug,
  'store=s'   => \$store,
@@ -33,6 +33,7 @@ GetOptions(
  'password=s'=> \$pass,
  'test=s'    => \$test,
  'acedb=s'   => \$ace_database,
+ 'dontclean' => \$dontclean,
 ) ||die(@!);
 
 if ($store) {
@@ -45,7 +46,7 @@ my $log = Log_files->make_build_log($wb);
 # clean out the database
 my $prefix = $wb->wormpep_prefix;
 my $pgdb = DBI->connect('dbi:Pg:dbname=clx;host=deskpro16391.dynamic.sanger.ac.uk',$user,$pass);
-$pgdb->do("DELETE FROM clustal WHERE peptide_id LIKE \'$prefix\%\'");
+$pgdb->do("DELETE FROM clustal WHERE peptide_id LIKE \'$prefix\%\'") unless $dontclean;
 
 # get wormpep dir and file name
 
@@ -64,7 +65,7 @@ $species ||= $wb->species;
 # submit jobs, one for every 5k of sequences
 while ($job <= ($count/5000)+1){
 	my $script="/software/worm/perl_510/bin/perl ~/wormbase/scripts/clustal_runner.pl -window 5000 -offset $job -user $user -pass $pass -pepfile $infile -database $acedb -species $species";
-	print `bsub -q long -e /dev/null -o /dev/null -J clw -R "select[mem>2000] rusage[mem=2000]" -M 2000000 $script`;
+	print `bsub -q long -e /dev/null -o /dev/null -J clw -R "select[mem>3000] rusage[mem=3000]" -M 3000000 $script`;
 	$job++;
 }
